[#1126] Consolidate window test mock setup
* Moves window mock setup into dedicated setupWindow.ts file
* Removes direct window state manipulation in test setup
* Improves test isolation by centralizing window mocking logic
* Switches from Window to OnetimeWindow type for better type safety

The change improves maintainability and reduces side effects in tests by
consolidating all window-related mock setup into a single location
