+++ tests/unit/vue/fixtures/window.fixture.ts
@@ -108,4 +108,4 @@ export const stateFixture: OnetimeWindow = {
 // Export the window fixture with the new structure
 export const windowFixture = {
   __ONETIME_STATE__: stateFixture,
-} as Window & typeof globalThis;
+} as OnetimeWindow & typeof globalThis;
